Fix delete api

+++ api/controllers/deleteAccount.js
@@ -3,6 +3,7 @@ const Accounts = require('../models/Accounts');
 
 module.exports = async (req, res, next) => {
   try {
+    if (!req.body.accountno) throw new ErrorHandler(400, 'Account Number is required');
     await Accounts.deleteOne({ accountno: req.body.accountno });
     res.json('Account Deleted Successfully!');
   } catch (err) {
